# swagger/openapi.yaml
openapi: 3.0.3
info:
  title: SLIIT AF — EcoMart API
  description: |
    REST API for the EcoMart e-commerce project (products, cart, orders, payments, reviews, admin).
    Use **Authorize** and paste `Bearer <token>` or only the JWT from login/register.
  version: 1.0.0
  contact:
    name: SLIIT AF Project
servers:
  - url: http://localhost:5000
    description: Local development
  - url: https://sliit-af-backend.onrender.com
    description: Production (Render)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT from POST /api/auth/login or /api/auth/register
  schemas:
...
    CreateOrderRequest:
      type: object
      required: [items, shippingAddress, phone]
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: string
              quantity:
                type: integer
                minimum: 1
        shippingAddress:
          type: string
        phone:
          type: string
        notes:
          type: string
        shippingLat:
          type: number
        shippingLng:
          type: number
...
    CartUpdateRequest:
      type: object
      properties:
        itemId:
          type: string
        quantity:
          type: integer
...
    ReviewBody:
      type: object
      properties:
        productId:
          type: string
        orderId:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
paths:
...
  /api/products:
    get:
      tags: [Products]
      summary: List products (filters via query — see controller)
      responses:
        '200':
          description: Product list
    post:
      tags: [Products]
      summary: Create product (multipart)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                title: { type: string }
                price: { type: number }
                description: { type: string }
                stock: { type: integer }
                productCategory: { type: string }
                certification: { type: string }
      responses:
        '201':
          description: Created
...
  /api/cart:
    get:
      tags: [Cart]
      summary: Get current user cart
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Cart
    put:
      tags: [Cart]
      summary: Update cart item quantity
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartUpdateRequest'
      responses:
        '200':
          description: Updated
...
  /api/orders:
    post:
      tags: [Orders]
      summary: Create order from cart
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created
...
  /api/admin/orders:
    get:
      tags: [AdminOrders]
      summary: All orders (paginated — query params per implementation)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Paginated orders
...
  /api/senara/reviews:
    get:
      tags: [Reviews]
      summary: All reviews (admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List
    post:
      tags: [Reviews]
      summary: Create review (customer)
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewBody'
      responses:
        '201':
          description: Created
...